Replace bitarrayCopy with a smaller, from-scratch version

+++ libuavcan/include/uavcan/marshal/bit_stream.hpp
@@ -21,8 +21,8 @@ namespace uavcan
  * @param dst_org       Destination array
  * @param dst_offset    Bit offset of the first destination byte
  */
-void bitarrayCopy(const unsigned char* src_org, unsigned src_offset, unsigned src_len,
-                  unsigned char* dst_org, unsigned dst_offset);
+void bitarrayCopy(const unsigned char* src_org, size_t src_offset, size_t src_len,
+                  unsigned char* dst_org, size_t dst_offset);
 
 /**
  * This class treats a chunk of memory as an array of bits.
